Compartmentalize IncidentCards components
